+++ docs/changelog.md
@@ -5,6 +5,7 @@ Next version
 ---
  * Drop support for php 7.4
  * Drop support for symfony 5.4
+ * Replace @Searchable and @SearchableField annotation by php8 attribute
 
 5.0
 ---